ml_service/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.0.3
sqlalchemy==2.0.30
python-dotenv==1.0.1
schedule==1.2.2
numpy==1.26.4
